chore: Move watchdog to separate source folder
