+++ shell/CMakeLists.txt
@@ -47,8 +47,8 @@ add_executable(${PROJECT_NAME}
         vsync/consumer_paced_vsync.cc
         timer.cc
         view/flutter_view.cc
-        watchdog.cc
-        stats.cc
+        watchdog/watchdog.cc
+        watchdog/stats.cc
         task_runner.cc
 )
