+++ shell/app.cc
@@ -37,7 +37,7 @@
 #include "view/flutter_view.h"
 
 #if BUILD_WATCHDOG
-#include "watchdog.h"
+#include "watchdog/watchdog.h"
 #endif
 
 #include <chrono>
